Add message filters.
Message filters are functions that get called when a message is
received. All message filter functions are called before the ddp message
is processed with the DDPClient.prototype._message method which is
guaranteed to be called last.

Message filters can be added in the constructor function options hash or
by calling the onMessage method with the function passed as a parameter.
The callback function takes two parameters: data and flags. Its context
(this) will be the DDPClient.