Fix ignored sends when server is disconnected
When the server is offline, a ddpclient.send would effectivley be
ignored. The callbacks would stil be registered, and the messages are
lost. The behavior is described
[here](https://github.com/vsivsi/meteor-job-collection/issues/203) in
more detail.

This patch makes send requests callback with errors if the server is
not connected. There is a small exception to the "connect" message,
which obviously shouldn't be blocked.